requirements.txt
colorama==0.4.6
matplotlib==3.8.3
numpy==2.0.0
pyconcorde==0.1.0
PyYAML==6.0.1
scipy==1.14.0
tabulate==0.8.10
torch==2.3.0
torch_geometric==2.5.3
torch_sparse==0.6.18
tqdm==4.66.4
tyro==0.8.4

# pip install git+https://github.com/facebookresearch/salina.git
# pip install pyconcorde